@@ -8,19 +8,27 @@
     "start": "node ./dist/App.js",
     "dev": "nodemon ./src/App.ts",
     "build": "tsc -p .",
-    "watchTS": "tsc -w --preserveWatchOutput "
+    "watchTS": "tsc -w --preserveWatchOutput ",
+    "up": "docker-compose -f docker-compose.db.yml up -d",
+    "stop": "docker-compose -f docker-compose.db.yml stop",
+    "down": "docker-compose -f docker-compose.db.yml down"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "@hapi/boom": "^9.1.0",
+    "@hapi/catbox": "^11.1.0",
+    "@hapi/catbox-redis": "^5.0.5",
+    "@hapi/cookie": "^11.0.1",
+    "@hapi/glue": "^7.0.0",
     "@hapi/hapi": "^19.1.1",
     "@hapi/joi": "^17.1.1",
     "bcrypt": "^4.0.1",
     "configue": "^1.3.4",
-    "db-migrate": "^0.11.10",
+    "db-migrate": "^0.11.11",
     "db-migrate-mysql": "^2.1.1",
+    "dotenv": "^8.2.0",
     "hapi-auth-jwt2": "^10.1.0",
     "inversify": "^5.0.1",
     "jsonwebtoken": "^8.5.1",
@@ -29,16 +37,20 @@
   },
   "devDependencies": {
     "@types/bcrypt": "^3.0.0",
-    "@types/bluebird": "^3.5.30",
+    "@types/bluebird": "^3.5.32",
     "@types/hapi-auth-jwt2": "^8.6.1",
+    "@types/hapi__catbox": "^10.2.2",
+    "@types/hapi__catbox-redis": "^5.0.0",
+    "@types/hapi__cookie": "^10.1.0",
+    "@types/hapi__glue": "^6.1.1",
     "@types/hapi__hapi": "^19.0.3",
     "@types/joi": "^14.3.4",
     "@types/jsonwebtoken": "^8.5.0",
-    "@types/mysql": "^2.15.10",
-    "@types/node": "^13.13.4",
+    "@types/mysql": "^2.15.12",
+    "@types/node": "^13.13.9",
     "@types/validator": "^13.0.0",
-    "nodemon": "^2.0.3",
-    "ts-node": "^8.9.1",
-    "typescript": "^3.8.3"
+    "nodemon": "^2.0.4",
+    "ts-node": "^8.10.1",
+    "typescript": "^3.9.3"
   }
 }
