@@ -16,7 +16,6 @@ import {
   updateUserSchema,
   updateUserPayloadSchema,
 } from "./UserSchemas";
-import { create } from "domain";
 
 @Controller("/users")
 @injectable()
@@ -51,6 +50,7 @@ class UserController {
   @Get("/{id?}")
   async get(request: Request): Promise<UserDto | Array<UserDto>> {
     try {
+      console.log("request", request);
       return request.params.id
         ? await this.userService.find(request.params.id)
         : await this.userService.findAll({ pageSize: 10 });
